Split out revert logic from forward logic in all options
